# Conflicts:
#	src/main/java/com/hmdandelion/project_1410002/common/exception/type/ExceptionCode.java
#	src/main/java/com/hmdandelion/project_1410002/sales/service/ClientService.java

+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/entity/order/OrderProduct.java
@@ -19,5 +19,5 @@ public class OrderProduct {
     @ManyToOne
     @JoinColumn(name = "order_code")
     private Order order;
-    private Integer productCode;
+    private Long productCode;
 }
